enc: move remote provider to new directory
