+++ src/enc/CMakeLists.txt
@@ -6,5 +6,4 @@ target_sources(enc PRIVATE enc.c)
 add_subdirectory(cipher)
 add_subdirectory(keyfmt)
 add_subdirectory(provider)
-add_subdirectory(qr)
 add_subdirectory(ui)
