+++ src/enc/provider/remote/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(enc PRIVATE remote.c)
+
+add_subdirectory(qr)
