+++ src/enc/provider/remote/qr/CMakeLists.txt
