+++ src/enc/provider/remote/qr/encqr.h
