+++ src/enc/provider/remote/qr/generate.c
