+++ src/enc/provider/remote/qr/lib.cpp
